+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed many typos.
 - Removed CODEOWNERS file to simplify repository management.
 - Enabled building for macOS.
+- Deregister users from push notifications on logout [#290](https://github.com/verse-pbc/issues/issues/290)
 
 ## [0.0.5]